requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
greenlet==3.2.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.10
nest-asyncio==1.6.0
openai==1.98.0
orjson==3.11.1
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==5.9.0
python-dotenv==1.1.1
python-telegram-bot==20.3
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.42
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
zstandard==0.23.0
gTTS
pydub
SpeechRecognition
aiohttp
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
apscheduler>=3.9.1
python-dateutil>=2.8.2
dateparser>=1.1.1  
python-multipart
orjson
pytest
pytest-asyncio
dateparser
fastapi
uvicorn
cryptography>=43.0.0
